requirements.txt
python-telegram-bot
google-generativeai
groq
requests
python-dotenv
pdfplumber
pydantic
pydantic-settings
